v4.7 güncellemesi.

+++ Presentation/Archieves.Kutuphane/ViewComponents/Comment/CommentListByBook.cs
@@ -10,7 +10,7 @@ namespace Archieves.Kutuphane.ViewComponents.Comment
         {
             _archievesService = archievesService;
         }
-        public IViewComponentResult Invoke(long id)
+        public IViewComponentResult Invoke(int id)
         {
             var values = _archievesService.GetCommentsAsync(id).Result.Value;
             if (values != null)
